Move gas time estimate utilities into utility file
